fix: package 명 typo 수정 applicaion -> application
