+++ src/main/java/nextstep/subway/application/PathService.java
@@ -1,6 +1,6 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.application;
 
-import nextstep.subway.applicaion.dto.PathResponse;
+import nextstep.subway.application.dto.PathResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Path;
 import nextstep.subway.domain.Station;
