+++ src/main/java/nextstep/subway/application/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import nextstep.subway.domain.Line;
