+++ src/main/java/nextstep/subway/application/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 public class StationRequest {
     private String name;
